Microservices

+++ demodatajpa/src/main/java/com/jay/demodatajpa/demodatajpa/entities/Processr.java
@@ -1,8 +1,11 @@
 package com.jay.demodatajpa.demodatajpa.entities;
 
+import java.util.List;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 
 import com.jay.demodatajpa.demodatajpa.dto.ProcessDTO;
 
@@ -12,7 +15,8 @@ public class Processr {
    private int no;
 	@Column(name = "model_name")
    private String modelName;
-   private double cost;
+   private int cost;
+   
    
    
 public int getNo() {
@@ -30,10 +34,10 @@ public void setModelName(String modelName) {
 public double getCost() {
 	return cost;
 }
-public void setCost(double cost) {
+public void setCost(int cost) {
 	this.cost = cost;
 }
-public Processr(int no, String modelName, double cost) {
+public Processr(int no, String modelName, int cost) {
 	super();
 	this.no = no;
 	this.modelName = modelName;
